Add websocket::stream timeouts
