+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 216:
 
 * Refactor websocket::stream operations
+* Add websocket::stream timeouts
 
 --------------------------------------------------------------------------------
