+++ test/beast/websocket/accept.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright (w) 2016-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
+// Copyright (c) 2016-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -10,6 +10,7 @@
 // Test that header file is self-contained.
 #include <boost/beast/websocket/stream.hpp>
 
+#include <boost/beast/_experimental/test/tcp.hpp>
 #include "test.hpp"
 
 #include <boost/asio/io_context.hpp>
@@ -614,6 +615,94 @@ public:
     }
 
     void
+    testTimeout()
+    {
+        using tcp = net::ip::tcp;
+
+        net::io_context ioc;
+
+        // success
+
+        {
+            stream<tcp::socket> ws1(ioc);
+            stream<tcp::socket> ws2(ioc);
+            test::connect(ws1.next_layer(), ws2.next_layer());
+
+            ws1.async_handshake("test", "/", test::success_handler());
+            ws2.async_accept(test::success_handler());
+            test::run_for(ioc, std::chrono::seconds(1));
+        }
+
+        {
+            stream<test::stream> ws1(ioc);
+            stream<test::stream> ws2(ioc);
+            test::connect(ws1.next_layer(), ws2.next_layer());
+
+            ws1.async_handshake("test", "/", test::success_handler());
+            ws2.async_accept(test::success_handler());
+            test::run_for(ioc, std::chrono::seconds(1));
+        }
+
+        // success, timeout enabled
+
+        {
+            stream<tcp::socket> ws1(ioc);
+            stream<tcp::socket> ws2(ioc);
+            test::connect(ws1.next_layer(), ws2.next_layer());
+
+            ws1.set_option(stream_base::timeout{
+                std::chrono::milliseconds(50),
+                stream_base::none(),
+                false});
+            ws1.async_accept(test::success_handler());
+            ws2.async_handshake("test", "/", test::success_handler());
+            test::run_for(ioc, std::chrono::seconds(1));
+        }
+
+        {
+            stream<test::stream> ws1(ioc);
+            stream<test::stream> ws2(ioc);
+            test::connect(ws1.next_layer(), ws2.next_layer());
+
+            ws1.set_option(stream_base::timeout{
+                std::chrono::milliseconds(50),
+                stream_base::none(),
+                false});
+            ws1.async_accept(test::success_handler());
+            ws2.async_handshake("test", "/", test::success_handler());
+            test::run_for(ioc, std::chrono::seconds(1));
+        }
+
+        // timeout
+
+        {
+            stream<tcp::socket> ws1(ioc);
+            stream<tcp::socket> ws2(ioc);
+            test::connect(ws1.next_layer(), ws2.next_layer());
+
+            ws1.set_option(stream_base::timeout{
+                std::chrono::milliseconds(50),
+                stream_base::none(),
+                false});
+            ws1.async_accept(test::fail_handler(beast::error::timeout));
+            test::run_for(ioc, std::chrono::seconds(1));
+        }
+
+        {
+            stream<test::stream> ws1(ioc);
+            stream<test::stream> ws2(ioc);
+            test::connect(ws1.next_layer(), ws2.next_layer());
+
+            ws1.set_option(stream_base::timeout{
+                std::chrono::milliseconds(50),
+                stream_base::none(),
+                false});
+            ws1.async_accept(test::fail_handler(beast::error::timeout));
+            test::run_for(ioc, std::chrono::seconds(1));
+        }
+    }
+
+    void
     testMoveOnly()
     {
         net::io_context ioc;
@@ -649,6 +738,7 @@ public:
     run() override
     {
         testAccept();
+        testTimeout();
         testMoveOnly();
         testAsioHandlerInvoke();
     }
